Removed mandatory prefixing of storage key.
Say I initialize the storage module without a prefix key:

    @NgModule({
      declarations: [
        AppComponent
      ],
      imports: [
        NgxLocalStorageModule.forRoot({ prefix: '' }),
      ],
    })
    export class AppModule {}

Keys will be prefixed regardless as "undefined_". With key "ponies":

  `${undefined}_${'ponies'}` -> undefined_ponies

If I supply a null or falsy prefixed value, what I expect to see is

  'ponies' -> ponies

This commit modifies storage service getters and setters in order to
only output a key prefix when it has a truthy value.